pyspark activations obj

+++ au/test/testconf.py
@@ -15,4 +15,5 @@ def use_tempdir(monkeypatch, test_tempdir):
   monkeypatch.setattr(conf, 'AU_TENSORBOARD_DIR', os.path.join(test_tempdir, 'tensorboard'))
   
   util.mkdir(test_tempdir)
-  util.rm_rf(test_tempdir)
+  if not os.environ.get('AU_NO_DEL_TEST_TEMPDIR'):
+    util.rm_rf(test_tempdir)
